+++ metro-extensions/contribute/impl-code-generators/src/test/kotlin/software/ralf/test/TestRobotGraph.kt
@@ -1,8 +1,0 @@
-package software.ralf.test
-
-import kotlin.reflect.KClass
-import software.ralf.app.platform.robot.Robot
-
-interface TestRobotGraph {
-  val robots: Map<KClass<*>, () -> Robot>
-}