+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/box/contributesrobot/defaultConstructorRobot.kt
@@ -4,14 +4,16 @@ import software.ralf.app.platform.inject.robot.ContributesRobot
 import software.ralf.app.platform.robot.Robot
 import software.ralf.app.platform.robot.RobotGraph
 
-@ContributesRobot(AppScope::class)
+abstract class TestScope private constructor()
+
+@ContributesRobot(TestScope::class)
 class TestRobot : Robot
 
-@DependencyGraph(AppScope::class)
-interface MyGraph : RobotGraph
+@DependencyGraph(TestScope::class)
+interface MyGraph
 
 fun box(): String {
-  val graph = createGraph<MyGraph>()
+  val graph = createGraph<MyGraph>() as RobotGraph
   val robotFactory = graph.robots.getValue(TestRobot::class)
   val robot = robotFactory()