+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/box/contributesrobot/injectClassSkipsProvider.kt
@@ -16,7 +16,7 @@ class TestRobot(
 ) : Robot
 
 @DependencyGraph(AppScope::class)
-interface MyGraph : RobotGraph
+interface MyGraph
 
 fun box(): String {
   val provider =
@@ -27,7 +27,7 @@ fun box(): String {
     return "FAIL: expected generated provider to be skipped"
   }
 
-  val graph = createGraph<MyGraph>()
+  val graph = createGraph<MyGraph>() as RobotGraph
   val robotFactory = graph.robots.getValue(TestRobot::class)
   val robot = robotFactory()