+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/box/contributesrobot/injectConstructorRobot.kt
@@ -16,10 +16,10 @@ class TestRobot(
 ) : Robot
 
 @DependencyGraph(AppScope::class)
-interface MyGraph : RobotGraph
+interface MyGraph
 
 fun box(): String {
-  val graph = createGraph<MyGraph>()
+  val graph = createGraph<MyGraph>() as RobotGraph
   val robotFactory = graph.robots.getValue(TestRobot::class)
   val robot = robotFactory()