+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/box/contributesrobot/injectSecondaryConstructorSkipsProvider.kt
@@ -18,7 +18,7 @@ class TestRobot private constructor(
 }
 
 @DependencyGraph(AppScope::class)
-interface MyGraph : RobotGraph
+interface MyGraph
 
 fun box(): String {
   val provider =
@@ -29,7 +29,7 @@ fun box(): String {
     return "FAIL: expected generated provider to be skipped"
   }
 
-  val graph = createGraph<MyGraph>()
+  val graph = createGraph<MyGraph>() as RobotGraph
   val robotFactory = graph.robots.getValue(TestRobot::class)
   val robot = robotFactory()