+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/diagnostics/contributesrobot/onlyAppScopeSupported.kt
@@ -1,8 +1,0 @@
-// RENDER_DIAGNOSTICS_FULL_TEXT
-package com.test
-
-import software.ralf.app.platform.inject.robot.ContributesRobot
-import software.ralf.app.platform.robot.Robot
-
-<!CONTRIBUTES_ROBOT_ERROR!>@ContributesRobot(String::class)<!>
-class TestRobot : Robot